Started working on profile page

+++ SportSync_BE/src/main/java/com/example/sportssync_be/repository/EntryRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface EntryRepository extends JpaRepository<Entry, Integer> {
+public interface EntryRepository extends JpaRepository<Entry, Long> {
 }
